requirements.txt
python-dotenv
streamlit
azure.core
azure-ai-documentintelligence
azure-storage-blob